Show all warnings + fix some. (#188)
* Show all warnings + fix some.

* Update protobuf_schema.cc

* Update CMakeLists.txt

* Fix warnings

* lint fix + warnings as errors on Windows

* Update CMakeLists.txt

* Update CMakeLists.txt

* Update protobuf_factory.cc

* Update protobuf_factory.cc

* Update protobuf_schema.cc

* Update CMakeLists.txt

* Update eth_helper_functions.h

* Update state_impl.cc

* more warnings fixes

* Update eth_helper_functions.h

* Update CMakeLists.txt

* Update status.h

* Update eth_helper_functions.h

* Update eth_helper_functions.h

* Update CMakeLists.txt

* disable pragma on non-clang compilers

* Update CMakeLists.txt

* Update io.cc

* Update CMakeLists.txt

+++ src/automaton/tests/crypto/test_SHA256_cryptopp.cc
@@ -76,5 +76,5 @@ TEST(SHA256_cryptopp, update_and_finish) {
 
 TEST(SHA256_cryptopp, digest_size) {
   SHA256_cryptopp hasher;
-  EXPECT_EQ(hasher.digest_size(), CryptoPP::SHA256::DIGESTSIZE);
+  EXPECT_EQ(hasher.digest_size(), static_cast<size_t>(CryptoPP::SHA256::DIGESTSIZE));
 }
